pyproject.toml
[tool.poetry]
name = "mushroom_classifier"
version = "0.1.0"
description = "A classifier for identifying mushroom species from images"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
torch = "^2.0.0"
torchvision = "^0.15.0"
pytorch-lightning = "^2.0.0"
hydra-core = "^1.3.0"
mlflow = "^2.0.0"
dvc = {extras = ["s3"], version = "^2.0.0"}
pillow = "^9.0.0"
numpy = "^1.20.0"
pandas = "^2.0.0"
scikit-learn = "^1.0.0"
onnx = "^1.13.0"
onnxruntime = "^1.14.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.3.0"
pytest = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88 